moving collection search to helper
