+++ models/location.js
@@ -1,7 +1,7 @@
 "use strict";
 var mongoose = require('mongoose');
+const routeCollectionHelper = require('../utils/routeCollectionHelper');
 var Action = require('./action');
-
 var Schema = mongoose.Schema;
 var ObjectId = Schema.ObjectId;
 
@@ -106,61 +106,6 @@ module.exports.toJsonList = function(item, descriptionMaxLength) {
     return ret;
 }
 
-module.exports.itemListToClientList = function(list, descriptionMaxLength, attachListToThisObject) {
-    return new Promise((resolve, reject) => {
-        if (!list) {
-            reject({ code: 400, message: "List is missing!" })
-        } else {
-            const arr = [];
-            let i, item;
-            for (i = 0; i < list.length; i++) {
-                item = Location.toJsonList(list[i], descriptionMaxLength);
-                arr.push(item);
-            }
-
-            attachListToThisObject.result = arr;
-            resolve(attachListToThisObject)
-
-        }
-    })
-}
-
-/**
- * List locations.
- *
- * @param {Object} query Query object to search for.  Pass null to list all locations.
- * @param {Number} maxDescriptionLength How long can Location.descriptions. pass null for no length restriction
- * @param {requestCallbackWithError} callback callback function where fist parameter is err and second is result
- */
-
 module.exports.search = function(query, sort, itemsPerPage, page, descriptionMaxLength) {
-    const sortedBy = sort ? sort : { lastModified: -1 };
-    page = Math.max(0, page);
-    return new Promise((resolve, reject) => {
-        Location
-            .find(query)
-            .limit(itemsPerPage)
-            .skip(itemsPerPage * page)
-            .sort(sortedBy)
-            .exec(function(err, list) {
-                if (err || !list) {
-                    reject(err)
-                } else {
-                    Location.countDocuments(query).exec(function(err, count) {
-                        if (err || !list) {
-                            reject(err);
-                        }
-                        const rootObject = {
-                            page: page,
-                            itemsPerPage: itemsPerPage,
-                            pages: Math.ceil(count / itemsPerPage),
-                            count: count
-                        }
-                        Location.itemListToClientList(list, descriptionMaxLength, rootObject)
-                            .then(newList => resolve(newList))
-                            .catch(err => reject(err));
-                    })
-                }
-            })
-    })
+    return routeCollectionHelper.collectionSearch('location', query, sort, itemsPerPage, page, descriptionMaxLength);
 };
